Add support for vim.xdl_diff
Requires PR neovim/neovim/#14536
