+++ .github/workflows/ci.yml
@@ -17,10 +17,14 @@ on:
 jobs:
   # This workflow contains a single job called "build"
   build:
+    strategy:
+      fail-fast: true
+      matrix:
+        neovim_branch: ['v0.5.0', 'nightly']
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
     env:
-      NEOVIM_BRANCH: v0.5.0
+      NEOVIM_BRANCH: ${{ matrix.neovim_branch }}
 
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
