+++ lua/gitsigns/actions.lua
@@ -361,7 +361,7 @@ M.preview_hunk = function()
    vim.cmd([[autocmd CursorMoved,CursorMovedI <buffer> ++once silent! unlet b:_gitsigns_preview_open]])
 
    if config.use_internal_diff then
-      local regions = require('gitsigns.diff_ffi').run_word_diff(hunk.lines)
+      local regions = require('gitsigns.diff_int').run_word_diff(hunk.lines)
       local offset = #lines - #hunk.lines
       for _, region in ipairs(regions) do
          local line, scol, ecol = region[1], region[3], region[4]
@@ -537,7 +537,7 @@ end
 
 local function run_diff(a, b)
    if config.use_internal_diff then
-      return require('gitsigns.diff_ffi').run_diff(a, b, config.diff_algorithm)
+      return require('gitsigns.diff_int').run_diff(a, b, config.diff_algorithm)
    else
       return require('gitsigns.diff_ext').run_diff(a, b, config.diff_algorithm)
    end
